fix: AdminSideBar no multi comp
