+++ app/frontend/src/components/admin/AdminSideBar.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import { SideBar } from '../layout/Sidebar';
+import { AdminItems } from './AdminItems';
+
+export function AdminSideBar({
+    children,
+    hideLogout = false,
+    hideLoginRegister = false,
+}: Readonly<{
+    children?: React.ReactNode;
+    hideLogout?: boolean;
+    hideLoginRegister?: boolean;
+}>) {
+  return <SideBar items={<AdminItems/>} hideLogout={hideLogout} hideLoginRegister={hideLoginRegister}>{children}</SideBar>
+}
