+++ app/frontend/src/pages/admin/live-map.tsx
@@ -1,4 +1,4 @@
-import { AdminSideBar } from '../../components/admin/adminSidebar';
+import { AdminSideBar } from '../../components/admin/AdminSideBar';
 import MapView from '../../components/guest/MapView';
 
 export default function AdminLiveMap() {
