+++ app/frontend/src/pages/admin/report-fire.tsx
@@ -1,4 +1,4 @@
-import { AdminSideBar } from '../../components/admin/adminSidebar';
+import { AdminSideBar } from '../../components/admin/AdminSideBar';
 import ReportPage from '../../components/reportfire/ReportPage';
 
 export default function AdminReportFire() {
